feat: enhance ONNX runtime availability handling and improve captions support
- Updated `runtime_status` function to reflect ONNX model availability based on the `captions` feature.
- Modified tests to verify ONNX availability aligns with the `captions` feature.
- Introduced a new timestamp formatting function for subtitles in `subtitles.rs`.
- Refactored `words.rs` to conditionally compile certain functions based on the `captions` feature.
- Updated tray menu event handling to properly manage async file opening.
- Changed app branding text in `app.html` for clarity.
- Enhanced `CaptionsPanel.svelte` to provide user feedback on remote transcription availability.
- Improved activity center visibility logic for export jobs in `activity-center.svelte`.
- Added backend profile management functions in `ipc.ts` and corresponding tests.
- Migrated profile management from localStorage to backend in `profiles.ts` and `profiles.svelte.ts`.
- Updated editor page to manage export panel visibility based on editor state.
- Added tests for panel logic to ensure correct functionality.
- Adjusted profile picker to handle async loading of profiles from the backend.

+++ apps/desktop/src-tauri/src/recording/mod.rs
@@ -336,7 +336,7 @@ fn resolve_region_target(rect: RegionRect) -> Result<CaptureTarget> {
             // The frontend Region Picker passes coordinates in PHYSICAL pixels,
             // but xcap's Monitor bounds are in LOGICAL pixels.
             // We must un-scale the physical center point to find the matching monitor.
-            let scale = display_scale_factor(monitor) as f32;
+            let scale = display_scale_factor(monitor);
             let cx = (center_x as f32 / scale).round() as i32;
             let cy = (center_y as f32 / scale).round() as i32;
 
@@ -349,7 +349,7 @@ fn resolve_region_target(rect: RegionRect) -> Result<CaptureTarget> {
         .context("unable to locate the display containing the selected region")?;
 
     let scale = display_scale_factor(&monitor);
-    let scale_f32 = scale as f32;
+    let scale_f32 = scale;
 
     // Convert the physical rect back to logical pixels so it shares the same
     // coordinate space as the monitor for clamping.
